Fixes for monospaced fonts
This adds two new functions
fix_monospaced_font which tries to fix the .null character that gets
added to truetype fonts with a width of 0 if it's the only character
that has a different width.

force_monospaced_font which currently is only applied to the
Nimbus Mono L font and which scales all characters so we force it to
be monospaced.