Add parameters to fix glyph unicode values
--fix-glyph-unicode : Try to fix unicode points and glyph names based
on glyph names containing hexadecimal codes (like "$0C00", "char12345"
or "uni004F")

--replace-unicode-values: When passed 2 comma separated numbers a,b
the glyph with an unicode value of `a` is replaced with the unicode value `b`.
Can be used more than once.

--shift-unicode-values: When passed 3 comma separated numbers a,b,c
this shifts the unicode values of glyphs between `a` and `b` (both included)
by adding `c`. Can be used more than once.